net: ethernet: prueth: rework node table code to use non-contiguous memory locations
On AM57x, the node/bin/index tables were all in one contiguous location
in memory, and this memory was allocated all at once, with one reference
to its base address. On AM3/4, these tables are in different,
non-contiguous locations, so now they will be allocated separately, with
a separate reference to each. The firmware revision used will determine
which locations/offsets to use.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>