TEMP: net: ethernet: prueth: workaround for AM3/4 node table collision
On AM3/4, index array is 64 for node table 128, so there is possible
collision. Linux driver checks flag set by firmware to see if a MAC from
SV frame is in the node table already, and in some cases this flag is
incorrectly set. This change removes the check, so ALL SV frames trigger
the insert function, it can be removed once that flag is always
correctly set.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>