net: ethernet: prueth: hsr/prp: fix debugfs path to have pruss id
Currently there are quirks added in the debugfs implementation to
satisfy the interface with snmpd prototype for hsr/prp. As this
series is cleaning this implementation to use proper interface for
this, cleanup the debugfs code to simply include the pruss-id in
the file path. Also fix some clean up issue in the current code.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>