hsr/prp: introduce net device lre ops for config and stats
LRE device requires configuration from user space through snmpd. This
patch introduces lredev ops in net_device struct to allow configure lre
parameters (attributes), query lre stats and node table from lower slave
devices when the lre is offloaded. This replaces the debugfs interface
files introduced at lower device driver for the same purpose as the
same can't be used in a production system. This API is similar to the
switch device ops and is implemented in a similar manner.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>