hsr/prp: add procfs interface for lredev config/stats
Currently it is not clear what interface to be used for accepting
configuration set/get or lre stats request or node table dump
request from user space such as that from snmpd. Currently
ip/udp/tcp mibs of mibii uses procfs for stats and such. So this
patch introduces the procfs interface at the upper layer hsr/prp
device files to allow lre configuration or display stats or node
table entries. procfs files are created under the
/proc/<interface-name> folder so that multiple hsr/prp
devices can co-exists. This patch removes the corresponding
files from debugfs.

The format of print to console is kept same as the current debugfs
output format expected by snmpd so that same text parser can be
re-used at snmpd even after this migration. It is expected that
this interface may not be acceptable for the upstream version of
the driver. So this has to be re-visited later.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>