hsr/prp: add link params to support VLAN tagged SV frames
In hsr/prp networks, user may choose to use vlan tagged SV frames
in the network which is defined as an optional feature in the
iec-62439 standard. User would then require to configure the VLAN ID,
PCP and CFI values to be used in the SV frames. This patch update the
driver to accept these values optionally from user as part of the
ip link add command and save the values in the priv structure of the
hsr/prp device.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>