hsr/prp: use vlan tag in the SV frame based on user choice
Update the driver to use the configured VLAN tag information
in the VLAN header of the SV frame. This essentially update
send_supervision_frame() to include vlan tag header and fill
the configured VLAN tag information in the header. Also update
is_supervision_frame() to consider VLAN header if present to
determine if a frame is a SV frame and frame_get_tagged_skb()
to check for vlan and use the correct offset for the SV payload.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>