iommu/omap: Add support for enabling DRA7xx system MMUs
The DRA7xx uses the system MMU1 for EDMA and MMU2 for PCIe. Routing
traffic through these MMUs is controlled with the CTRL_CORE_SMA_SW_7
register bit field.

The devicetree property "ti,syscon-mmuconfig" is already used for enabling
the MMUs inside the DSP subsystems. Here, a subsystem id argument (0/1)
determines the value for the corresponding control register.

A more general approach is implemented for enabling the system MMUs:
Instead of using an id, an offset argument is used to obtain the register
address within the syscon node. An additional argument specifies the bit
field value for enabling the MMU.

Signed-off-by: Ralf Goebel <ralf.goebel@imago-technologies.com>