Documentation: dt: Update OMAP iommu bindings for DRA7xx system MMUs
The DRA7xx uses the system MMU1 for EDMA and MMU2 for PCIe. Routing
traffic through these MMUs is controlled with the CTRL_CORE_SMA_SW_7
register bit field.

The existing syscon property "ti,syscon-mmuconfig" is modified to
also work with the system MMUs.

An example for using the PCIe MMU is also added.

Signed-off-by: Ralf Goebel <ralf.goebel@imago-technologies.com>