iommu/omap: Add support for storing page tables within SRAM
Page tables are normally placed in SDRAM which causes latencies when reading
page table entries while the SDRAM controller is busy with other requests.

This patch enables support for placing the page tables into the SoC's SRAM
in order to reduce the latency for external DMA masters. The implementation
uses the genalloc/genpool subsystem for SRAM management.

Signed-off-by: Ralf Goebel <ralf.goebel@imago-technologies.com>